mcp-servers/deep-biblio/pyproject.toml
[project]
name = "deep-biblio-mcp-server"
version = "0.1.0"
description = "MCP server for deep-biblio-tools integration"
authors = [
    {name = "Petteri Teikari", email = "petteri.teikari@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "loguru>=0.7.0",
    # deep-biblio-tools dependencies
    "requests>=2.25.1",
    "beautifulsoup4>=4.9.3",
    "bibtexparser>=1.4.0",
    "lxml>=4.6.3",
    "pypdf2>=3.0.1",
    "pdfplumber>=0.11.5",
    "click>=8.0.0",
    "tqdm>=4.65.0",
    "pypandoc>=1.11",
    "python-dotenv>=1.0.0",
    "markdown-it-py>=3.0.0",
    "mistletoe>=1.4.0",
    "pylatexenc>=2.10",
    "pyyaml>=6.0.2",
    "pyzotero>=1.7.3",
]

[project.scripts]
deep-biblio = "deep_biblio.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deep_biblio"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.15.1",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
